@@ -5,15 +5,17 @@
   "dependencies": {
     "@dhaiwat10/react-link-preview": "^1.14.2",
     "@reduxjs/toolkit": "^1.8.3",
+    "@tanstack/react-query": "^4.6.0",
+    "@tanstack/react-query-devtools": "^4.6.0",
     "@tanstack/react-virtual": "^3.0.0-alpha.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.15",
     "@types/lodash": "^4.14.184",
-    "@types/node": "^17.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/node": "^18.7.15",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
     "@types/react-icons": "^3.0.0",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
@@ -35,8 +37,8 @@
     "i18next-browser-languagedetector": "^6.1.2",
     "i18next-xhr-backend": "^3.2.2",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet-async": "^1.2.2",
     "react-i18next": "^11.15.0",
     "react-icons": "^4.4.0",
@@ -56,7 +58,7 @@
     "react-virtualized-auto-sizer": "^1.0.6",
     "react-window": "^1.8.7",
     "react-window-infinite-loader": "^1.0.8",
-    "recoil": "^0.5.2",
+    "recoil": "^0.7.5",
     "rxjs": "^7.5.6",
     "socket.io-client": "^4.4.0",
     "styled-components": "^5.3.5",
@@ -97,6 +99,7 @@
   "devDependencies": {
     "@commitlint/cli": "^15.0.0",
     "@commitlint/config-conventional": "^15.0.0",
+    "babel-loader": "8.1.0",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
     "lint-staged": "^12.1.2",
