@@ -1,3 +1,4 @@
+import { StyledContainer, StyledFlex } from '@components/shared/shared-style';
 import React from 'react';
 import styled from 'styled-components';
 
@@ -8,23 +9,17 @@ type Props = {
 
 export const OneSideBarLayout = ({ sideBar, content }: Props) => {
   return (
-    <Container>
-      <SideBar>{sideBar}</SideBar>
-      <MainContent>{content}</MainContent>
-    </Container>
+    <StyledContainer>
+      <StyledFlex gap={2.5}>
+        <SideBar>{sideBar}</SideBar>
+        <MainContent>{content}</MainContent>
+      </StyledFlex>
+    </StyledContainer>
   );
 };
 
-const Container = styled.div`
-  max-width: 75%;
-  flex: 1;
-  margin: 2.5rem auto;
-  display: flex;
-  gap: 2.5rem;
-  align-items: flex-start;
-`;
 const SideBar = styled.aside`
-  width: 30%;
+  width: 25%;
 `;
 const MainContent = styled.div`
   flex: 1;
