@@ -7,9 +7,10 @@ import { EUserQuery, LONG_STATE_TIME } from '@constants';
 import LayoutWithHeader from '@layout/layout-with-header';
 import { OneSideBarLayout } from '@layout/one-sidebar.layout';
 import { IUser } from '@type/user.type';
-import React, { useEffect, useMemo, useState } from 'react';
+import { queryStringToObject } from '@utils/helper';
+import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useQuery } from 'react-query';
+import { useQuery } from '@tanstack/react-query';
 import { useLocation, useNavigate, useParams } from 'react-router';
 import { EventBusName, onPushEventBus } from 'services/event-bus';
 import useUserService from 'services/user.service';
@@ -18,33 +19,30 @@ import ProfileHomePage from './profile-home';
 import ProfileLikedTweetPage from './profile-liked-tweet';
 import ProfileMediaPage from './profile-media';
 
-enum EProfileScreen {
+export enum EProfileScreen {
   Home = 'home',
   Liked = 'liked',
   Medias = 'medias',
 }
 
 const ProfilePage = () => {
-  const { getUser } = useUserService();
+  const { getUser, getCurrentUser } = useUserService();
   const location = useLocation();
-  const { userId } = useParams();
+  const params = useParams();
+  const userId = params.userId;
   const { t } = useTranslation();
+  const navigate = useNavigate();
 
-  const [screen, setScreen] = useState<EProfileScreen>(EProfileScreen.Home);
-
-  useEffect(() => {
-    console.log('location', location);
-  }, [location.pathname]);
+  const { screen } = queryStringToObject(location.search);
 
   const onChangeScreen = (nextScreen: EProfileScreen) => {
     if (screen !== nextScreen) {
-      // set pathname to url
-      location.pathname = `profile/${nextScreen}/${userId}`;
-      setScreen(nextScreen);
+      navigate({
+        pathname: location.pathname,
+        search: `?screen=${nextScreen}`,
+      });
     }
   };
-
-  const navigate = useNavigate();
 
   const { data: userData, isLoading } = useQuery<IUser>(
     [EUserQuery.GetUser, userId],
@@ -94,6 +92,8 @@ const ProfilePage = () => {
       },
     ];
   }, [t]);
+
+  console.log('userData', userData);
 
   return (
     <React.Fragment>
